feat(build): add initial build script
add a simple build script that sets up a gopath and uses the current git
directory for the github.com/coreos/etcd packages.

There aren't a lot of great alternatives to doing it this way unless we
want to check in all of the dependencies into the repo (which is
actually a good practice probably).