feat(watch/stream): add "stream" param for "watch" requests
This commit adds support for streaming watch requests.

Normally when waiting for a change, the response ends after writing
the first change. With this commit, if you pass "stream=true" in the
query parameters the response is not closed after writing the first
change to the response. Instead, the connection remains open and
subsequent changes are written to the response body until the connection
is closed by the requester. Single newline charachters are printed
between each changes.

This feature is useful in scenarios that you need to watch too many
changes continuosly. It is faster than making a new HTTP connection for
each watch.