hack(store/store): multiple sets on a directory
We should allow people to set the ttl on a directory multiple times in a
row. As it is you have to first do PUT foobar2?ttl=30 and then PUT
foobar2?ttl=30?prevExist=true

$ ./bin/etcdctl setdir foobar2  --ttl 30
$ ./bin/etcdctl setdir foobar2  --ttl 30
Error: 102: Not a file (/foobar2) [2]

Reported by AmandaC in IRC.

This patch is not mergeable but we need to figure out something that is.