Update clients-matrix.md
Swapped the clients with the features on the clients matrix so it is more comprehensible. The old version got too large to show on the screen. This new layout added a language column so you can quickly see the language of the client library.

Also moved v1-only clients to the bottom since v2 is what most people are looking for today.