etcdserver: initialise ServerStats with leader name if available
When starting an EtcdServer set leader name in ServerStats if a leader already defined in LeaderStats. On a single machine cluster there are no AppendRequest calls to set leader name, while the machine is an active leader stats shows a default string as leader name. With this fix the name is correctly set.

Fixes #561