etcdserver: begin support for providing a custom Logger
When embedding etcd in another app it would be helpful to have control
over the server's logging output. The higher level app can then squelch
all output unless verbose logging is requested or redirect it to a
special log file.

Fatal errors are left alone to ensure the error is always reported when
the app crashes.

This change is incomplete, just submitting for feedback.