raft: attach commit index into the heartbeat response
Follower attaches commit index in heartbeat message. This is a preparation for
not resending snapshot until the pending one is failed. There is a case that
the snapshot is received sucessfully but the response is lost. The only active
channel between the leader and the follower at this point is the heartbeat, since
the snapshot resending is disabled. We need to attach some information into the
heartbeat to ack the leader that the snapshot is actually received.