raft: add progress CatchUp
CatchUp marks whether the remote node catches up with the leader. If it
does, leader can send MsgApp optimistically to the follower.

It comes into CatchUp status when leader receives a valid MsgAppResp
from the follower.

When it is bootstrapped or rejected by the remote, CatchUp is set to
false, and it falls back to slow replication method.