etcdserver: omit updateattr on non-exist member in Cluster
The omission happens only when newly joined member catches up the
progress.

Its Cluster struct has a future view of members, which doesn't have
members that are removed between local index and future index.

The non-exist member should be removed from the cluster in further apply,
so it is ok to omit the updateattr in Cluster.