Retry on rejected empty appends
In the event of an empty append (used to advance the commit index)
being rejected, the follower will appear to be caught up despite the
commit index not having been advanced. In our experience, this can
happen when replies arrive out of order.

This commit adds a clause to handle that edge case by resending the
append.

This was discovered when running the github.com/cockroachdb/cockroach
test suite under go1.5beta1.