Dockerfile: use a consistent default user id
Docker defaults to root. We really would rather etcd not run as root.
For lack of a better choice lets use 2379, the etcd port number, as the
default.

We should probably start documenting this elsewhere.

WARNING: when we merge this we will need to warn folks that upgrading
from version X to Y will require they chown their data directories.
I suggest we hold off on merging this until 2.3.