rafthttp: support sending v3 snapshot
After sending raft snapshot message as before, sender will wait for
receiver to request the full snapshot and send it to the receiver.
Only after sender sends out the full snapshot, it thinks that snapshot
is sent successfully.
Only after receiver receives the full snapshot, it acknowledges the
snapshot is received and process the raft snapshot message.