proxy: shorter refresh interval for initial set-up
This is a hacky way to expedite some proxy refresh for https://github.com/coreos/etcd/issues/3647.
The current default proxy refresh interval is 30 seconds. If the director has
no endpoint list yet, it has to wait 30 seconds to refresh. For the initial
iterations, we can shorten the intervals to decrease the wait time.