scripts: cd into correct directory in genproto.sh
This is a workaround for https://github.com/coreos/etcd/issues/3815.
genproto.sh script is not working when run in scripts directory because it
executes the command pushd assuming that the working directory is the project
root directory (github.com/coreos/etcd). This fix assumes that the script would
be run IN the scripts directory and cds into the parent directory when it needs
to run pushd.