client: Fix incorrect handling of context errors
If a client request fails due to a context error (either a timeout or
cancellation), the client code is supposed to return that error
directly.  However, before this patch, Client was returning
ErrClusterUnavailable due to a big detecting context errors.  This
patch fixes the bug so that the appropriate context error is returned
instead.