client: new envvar ETCDCLIENT_PEERS for supplying peer URLs
Currently, programs using etcd client library need to supply peer
URLs via client.Config.Endpoints. It requires a mechanism for setting
the URLs to the variable via command line options, config files, or
environment variables in their own ways. It would be a little bit
confusing especially in a case that multiple client programs depend on
a single etcd cluster.

For avoiding the problem, this commit introduces a new environment
variable ETCDCLIENT_PEERS for a standard way of supplying the URLs.