client: an option for prioritizing a leader node as a request destination
Current etcd client library chooses a default destination node from
every member of a cluster in a random manner. However, requests of
write and read (for consistent results) need to be forwarded to the
leader node as the nature of Raft algorithm. If the chosen node is a
follower, additional network traffic will be caused by the forwarding
from follower to leader.

For reducing the forward traffic, this commit adds a new option
PrioritizeLeader to client.Config. If this variable is true, the
library chooses the leader node as a default destination of request.

I evaluated the effectiveness of this commit with 4 t1.micro
instances of AWS (3 nodes for etcd cluster and 1 node for etcd
client). Client executes this simple benchmark
(https://github.com/mitake/etcd-things/tree/master/prioritize-leader-bench),
just writes 10000 keys. When PrioritizeLeader is false, the benchmark
needed 1 min and 32.102 sec to finish. When it is true, the benchmark
needed 1 min 4.760 sec.