client: add a new endpoint selection mode for prioritizing follower
This commit adds a new endpoint selection mode for prioritizing
follower, EndpointSelectionPrioritizeFollower. If this mode is used,
followers are prioritized as request destination.

This is just for benchmarking purpose. Choosing a leader or a follower
as a request destination produces very different scores in typical
benchmark. For measuring worse score for certain, the new mode is
useful.