etcdmain: Refactor etcdStart for use in embedding etcd servers
This patch refactors out the important (and often non-public) portions
of etcdStart that need to get run for use in embedding an etcd server.
This will allow projects just as `mgmt` (config) to embed etcd without
duplicating code, or trailing behind the latest upstream etcd changes.
It puts all of the common network connection code into a new struct...