etcdserver: improved discard of previous uncommitted entries
The earler processing was matching entries for index in a for loop which
would have taken longer if commited entires present were a lot more and
led to unnecessary taking more time for removing uncommited entries.
For example if st.commit = 100 and entries present in ents from index 1
to 105 then the for loop will run 101 times before it removes the 5
uncommited entries.