*: a new option for configuring a coefficient of randomized election timeout
This commit adds a new option --randomized-election-coefficient. The
value specified with the option will be used for calculating the
randomized election timeout in
raft.resetRandomizedElectionTimeout(). If the coefficient is low
(e.g. 0), randomness of the election will become low and split
vote will likely happen.

Of course, the split vote decreases availability and nobody wants
it. However, if we can cause the situation intentionally, it will be
useful for debugging the candidate state that is ephemeral in
realistic deployments and hard to be reproduced in testing
environments.

Actually, letting the coefficient 0 in testing on a single machine
(just adding --randomized-election-coefficient 0.0 to entries of
Procfile) shows lots of split vote situation after killing a leader
node.