wal: only consider a wal directory as Existing if it has wal files
In cases where etcd is started with a user-defined wal directory, it's possible
there may already be contents in the directory (e.g., lost+found on a fresh
filesystem), which will be treated as initialized even though it is not.
Instead, if there are no wal files, the directory should be treated as an
uninitialized wal directory.