raft: Remove return error from raft.Step
`raft.Step` never returned an error, but yet had an error in its
signature, which was in turn propagated in a number of places. This
change gets rid of the unnecessary return value.

I suspect that this may have been intentional in order to maintain some
form of backwards-compatibility somewhere.